Initial Release
The method of interest is - uniqueDeviceIdentifier. This method,
depending on the value of kShouldAlwaysGenerateCustomUDID (in the header
file) will return a custom UDID if iOS 5 is detected or Apple's UDID if
a pre iOS 5 OS is detected. Users can customize this behavior by
changing kShouldAlwaysGenerateCustomUDID to 1, which will result in the
class always outputting a custom UDID.